04-email-triage-agent/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
langfuse
